[plugins] allow calling of hooks on a per-plugin basis
Previously, when sending a request to a `server.api.plugin-request`
hook (think webhook), instead of calling the handler function for
the plugin specified in the URL, we called *all* handler functions
for all plugins for this hook.

This fixes that and only calls the hook that was requested.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:
http://go.hood.ie/thehoodiefirm