refactor: stop sending info to process
BREAKING CHANGE: Before hoodie-server was sending a subset of its configuration
to the process. The hoodie-cli could pick that up and log it out nicely. Now
is no longer started via the CLI and thus this is removed. It's breaking
nonetheless, because the CLI can no longer work with it.

* * *

This commit was sponsored by The Hoodie Firm.
You can hire The Hoodie Firm:

http://go.hood.ie/thehoodiefirm