refactor(store): use `server.plugins.store.api`
BREAKING CHANGE:

The API now keeps state of all databases and access in a dedicated
`hoodie-store` database. If your app uses CouchDB as its backend, there
is no migration required, it happens automatically on next restart.

If you don’t use CouchDB (which is the default), the simplest way to
"migrate" is to delete the `.hoodie/data` folder. It will loose all
data, so this is only an option for local development. If you don’t want
to loose all data, you manually have to create the `hoodie-store`
database with documents for each user database. The documents look
like this:

```json
{
   "_id": "db_user/lsdlo55",
   "_rev": "1-a01e6998fb10543ca0402e648dd1d048",
   "access": {
       "read": {
           "role": [
               "id:lsdlo55"
           ]
       },
       "write": {
           "role": [
               "id:lsdlo55"
           ]
       }
   }
}
```

Besides that, this release also updates to a new express-pouchdb version
which includes a bugfix for a faulty encoding of database folders. You
must rename all folders that look like `user%2Fabc4567` to
`user/abc4567` (nested directories). See pouchdb/express-pouchdb#377