fix: do not set `authenticationDb` name from CouchDB config
BREAKING CHANGE:

CouchDB’s `_users` database is configurable at `/_config/couch_httpd_auth/authentication_db`. We read out this configuration and used it as the name of the users database. But as we don’t rely on the special behavior of CouchDB’s users database (auto-hashing of passwords) we can use any normal CouchDB database to store the accoutns of Hoodie. And as the `/_config` API is not implemented by CouchDB-like hosters like Cloudant, we want to store this configuration ourselves.

For now, we hardcode the database name to default to `_users`. The name can be configured via `options.account.usersDb`.