feat: adminPassword option
BREAKING CHANGE:

Before this change, we loaded admin accounts from CouchDB’s `/_config/admins` API or for PouchDB we simply set the password to `"secret"` and wrote it into `.hoodie/config.json`

We dropped persisting the admin password altogether and instead made it an option: `adminPassword`. If it’s not set, then no admin account exists.